pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "dbms"
version = "0.1.0"
description = "A custom DBMS implementation"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "GswapG", email = "swappyontube@gmail.com"},
]
dependencies = [
    "PyYAML",
    "ply"
]

[project.optional-dependencies]
dev = [
    "types-PyYAML"
]

[tool.setuptools]
packages = ["dbms"]
package-dir = {"dbms" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/integration"]
python_files = ["test_*.py"]
addopts = "--cov=dbms --cov-report=term-missing"

[tool.coverage.run]
source = ["dbms"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
strict_optional = true

# Paths to check
files = ["src"]
exclude = ["tests", "parsetab.py", "src/parser/parsetab.py"]

# Configure import discovery
ignore_missing_imports = true

# Be strict with source code
[[tool.mypy.overrides]]
module = "dbms.*"
strict = true

# Ignore tests completely
[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.poetry.dependencies]
graphviz = "*"
